+++ EventSourcing.Core.Tests/MockAggregates/EmptyEvent.cs
@@ -1,0 +1,4 @@
+namespace EventSourcing.Core.Tests.MockAggregates
+{
+    public class EmptyEvent : Event { }
+}
