+++ EventSourcing.Core.Tests/MockAggregates/SimpleAggregate.cs
@@ -1,4 +1,4 @@
-namespace EventSourcing.Core.Tests.Mocks
+namespace EventSourcing.Core.Tests.MockAggregates
 {
     internal class SimpleAggregate : Aggregate
     {
