--- EventSourcing.Core.Tests/Mocks/EmptyEvent.cs
@@ -1,4 +1,0 @@
-namespace EventSourcing.Core.Tests.Mocks
-{
-    public class EmptyEvent : Event { }
-}
